backend/requirements.txt
# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis & Caching
redis==5.0.1
aioredis==2.0.1

# Task Queue
celery==5.3.4
celery[redis]==5.3.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1

# ML & Pattern Discovery
tigramite==5.2.4.1  # PCMCI causal discovery
stumpy==1.12.0  # Matrix profile for recurring patterns
mlxtend==0.23.0  # Association rules mining
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4
scikit-learn==1.3.2

# HuggingFace Models (MOMENT)
# Note: momentfm is experimental - uncomment when ready for MVP2
# momentfm==0.1.0
# torch==2.1.1
# transformers==4.35.2

# Data Processing
pandas==2.1.3
numpy==1.26.2
python-dateutil==2.8.2
pytz==2023.3

# HTTP Client (for external API calls)
httpx==0.25.2
aiohttp==3.9.1

# Validation & Serialization
pydantic==2.5.0
email-validator==2.1.0

# Environment Variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing FastAPI

# Code Quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

# Monitoring & Logging
structlog==23.2.0
python-json-logger==2.0.7

# CORS
fastapi-cors==0.0.6

# Health Checks
psutil==5.9.6

# Utilities
tenacity==8.2.3  # Retry logic
